// types/enums.d.ts
export type Gender = "MALE" | "FEMALE" | "NONE" // 상황에 맞게 추가
export type UserStatus = "ACTIVE" | "INACTIVE" | "SUSPENDED" // 상황에 맞게 추가
export type UserRole = "ROLE_USER" | "ROLE_ADMIN" // 상황에 맞게 추가
export type GoalType = "strength" | "hypertrophy" | "fatLoss" | "recomposition" | "generalFitness"
export type SplitType = "fullBody" | "upperLower" | "pushPullLegs" | "bodyPartSplit" | "custom"
export type ExperienceLevel = "beginner" | "intermediate" | "advanced"